adding camer engine and manager headers
