+++ app/src/main/cpp/CMakeLists.txt
@@ -4,6 +4,7 @@
 
 cmake_minimum_required(VERSION 3.4.1)
 
+set(COMMON_SOURCE_DIR ${CMAKE_CURRENT_SOURCE_DIR}/../../../../common)
 
 
 # Specifies a library name, specifies whether the library is STATIC or
@@ -16,9 +17,11 @@ cmake_minimum_required(VERSION 3.4.1)
 #Add a library to the project using the specified source files.
 add_library(gl_code SHARED
         gl_code.cpp)
+
 # build native_app_glue as a static lib
 include_directories(${ANDROID_NDK}/sources/android/native_app_glue
         ${COMMON_SOURCE_DIR})
+
 
 add_library(app_glue STATIC
         ${ANDROID_NDK}/sources/android/native_app_glue/android_native_app_glue.c)
@@ -57,6 +60,7 @@ target_link_libraries(understand_triangle
         log
         EGL
         GLESv2)
+
 target_link_libraries(camera_stream
         android
         log
