+++ app/src/main/cpp/camera_manager.h
@@ -1,0 +1,13 @@
+//
+// Created by 14035 on 2024-11-08.
+//
+
+#ifndef CLICKER_CAMERA_MANAGER_H
+#define CLICKER_CAMERA_MANAGER_H
+
+#endif //CLICKER_CAMERA_MANAGER_H
+
+
+class NDKCamera {
+
+};
